refactor: split ble transport from device
